feat: refactor map components into separate files and improve code organization
